refactor folder structure and add router to it
