+++ src/ts/components/header.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function Header() {
+  return <h1>Header</h1>
+}
